fix: Filter influencing traits to lookup matching kits
